pyproject.toml
[project]
name = "rl-transaction-attack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<4"
dependencies = [
    "deprecated>=1.2.18",
    "imblearn>=0.0",
    "jupyterlab>=4.4.2",
    "matplotlib>=3.9.2",
    "multi-agent-rlenv>=3.3.7",
    "optuna-dashboard>=0.18.0",
    "optuna>=4.3.0",
    "orjson>=3.10.18",
    "pandas>=2.2.3",
    "polars>=1.14.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "scikit-learn>=1.5.2",
    "scipy>=1.15.2",
    "seaborn>=0.13.2",
    "torch>=2.5.1",
    "tqdm>=4.67.1",
    "typed-argparse>=0.3.1",
    "gunicorn>=23.0.0",
    "ipywidgets>=8.1.7",
    "pyarrow>=20.0.0",
    "pytest>=8.3.5",
]

[tool.uv.sources]
torch = [{ index = "pytorch-cu118" }]


[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true


[dependency-groups]
dev = ["ruff>=0.8.0"]


[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = "src"
